+++ packages/maker/src/api/defineAction.ts
@@ -153,7 +153,7 @@ export function defineCreateInterfaceObjectAction(
 export function defineCreateObjectAction(
   def: ActionTypeUserDefinition,
 ): ActionType {
-  Object.keys(def.parameterLevelValidations ?? {}).forEach(id => {
+  Object.keys(def.parameterConfiguration ?? {}).forEach(id => {
     invariant(
       def.objectType.properties?.[id] !== undefined,
       `Property ${id} does not exist on ${def.objectType.apiName}`,
@@ -178,14 +178,15 @@ export function defineCreateObjectAction(
         displayName: def.objectType.properties?.[id].displayName
           ?? convertToDisplayName(id),
         type: extractActionParameterType(def.objectType.properties?.[id]!),
-        validation: (def.parameterLevelValidations?.[id] !== undefined)
+        validation: (def.parameterConfiguration?.[id] !== undefined)
           ? {
-            ...def.parameterLevelValidations?.[id],
-            allowedValues: def.parameterLevelValidations?.[id].allowedValues
+            ...def.parameterConfiguration?.[id],
+            allowedValues: def.parameterConfiguration?.[id].allowedValues
               ?? extractAllowedValuesFromType(
                 def.objectType.properties?.[id].type!,
               ),
-            required: def.parameterLevelValidations?.[id].required ?? true,
+            required: def.parameterConfiguration?.[id].required ?? true,
+            defaultValue: def.parameterConfiguration?.[id].defaultValue,
           }
           : {
             required: (def.objectType.properties?.[id].array ?? false)
@@ -237,6 +238,10 @@ export function defineCreateObjectAction(
         ],
       }
       : {}),
+    ...(def.defaultFormat && { defaultFormat: def.defaultFormat }),
+    ...(def.enableLayoutSwitch
+      && { enableLayoutSwitch: def.enableLayoutSwitch }),
+    ...(def.displayAndFormat && { displayAndFormat: def.displayAndFormat }),
   });
 }
 
@@ -341,7 +346,7 @@ export function defineModifyInterfaceObjectAction(
 export function defineModifyObjectAction(
   def: ActionTypeUserDefinition,
 ): ActionType {
-  Object.keys(def.parameterLevelValidations ?? {}).forEach(id => {
+  Object.keys(def.parameterConfiguration ?? {}).forEach(id => {
     invariant(
       def.objectType.properties?.[id] !== undefined,
       `Property ${id} does not exist on ${def.objectType.apiName}`,
@@ -367,14 +372,15 @@ export function defineModifyObjectAction(
         displayName: def.objectType.properties?.[id].displayName
           ?? convertToDisplayName(id),
         type: extractActionParameterType(def.objectType.properties?.[id]!),
-        validation: (def.parameterLevelValidations?.[id] !== undefined)
+        validation: (def.parameterConfiguration?.[id] !== undefined)
           ? {
-            ...def.parameterLevelValidations?.[id],
-            allowedValues: def.parameterLevelValidations?.[id].allowedValues
+            ...def.parameterConfiguration?.[id],
+            allowedValues: def.parameterConfiguration?.[id].allowedValues
               ?? extractAllowedValuesFromType(
                 def.objectType.properties?.[id].type!,
               ),
-            required: def.parameterLevelValidations?.[id].required ?? false,
+            required: def.parameterConfiguration?.[id].required ?? false,
+            defaultValue: def.parameterConfiguration?.[id].defaultValue,
           }
           : {
             required: (def.objectType.properties?.[id].array ?? false)
@@ -439,6 +445,10 @@ export function defineModifyObjectAction(
         ],
       }
       : {}),
+    ...(def.defaultFormat && { defaultFormat: def.defaultFormat }),
+    ...(def.enableLayoutSwitch
+      && { enableLayoutSwitch: def.enableLayoutSwitch }),
+    ...(def.displayAndFormat && { displayAndFormat: def.displayAndFormat }),
   });
 }
 
@@ -699,6 +709,8 @@ function extractAllowedValuesFromType(
           return type.markingType === "CBAC"
             ? { type: "cbacMarking" }
             : { type: "mandatoryMarking" };
+        case "string":
+          return { type: "text" };
         case "struct":
           throw new Error("Structs are not supported yet");
         default:
@@ -719,6 +731,8 @@ function extractActionParameterType(
     switch (typeType.type) {
       case "marking":
         return maybeAddList("marking", pt);
+      case "string":
+        return maybeAddList("string", pt);
       case "struct":
         throw new Error("Structs are not supported yet");
       default:
@@ -832,6 +846,12 @@ function validateActionValidation(action: ActionType): void {
         action.parameters,
       );
     });
+    if (param.validation.defaultValue?.type === "staticValue") {
+      invariant(
+        param.validation.defaultValue.staticValue.type === param.type,
+        `Default static value for parameter ${param.id} does not match type`,
+      );
+    }
     seenParameterIds.add(param.id);
   });
 }