+++ packages/create-app.template.tutorial-todo-app/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @osdk/create-app.template.tutorial-todo-app
 
+## 2.4.0-beta.9
+
+## 2.4.0-beta.8
+
 ## 2.4.0-beta.7
 
 ## 2.4.0-beta.6