+++ packages/foundry-sdk-generator/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @osdk/foundry-sdk-generator
 
+## 2.4.0-beta.9
+
+### Patch Changes
+
+- Updated dependencies [9101bad]
+  - @osdk/client@2.4.0-beta.9
+  - @osdk/api@2.4.0-beta.9
+  - @osdk/generator@2.4.0-beta.9
+
+## 2.4.0-beta.8
+
+### Patch Changes
+
+- Updated dependencies [9f4fe9e]
+  - @osdk/client@2.4.0-beta.8
+  - @osdk/api@2.4.0-beta.8
+  - @osdk/generator@2.4.0-beta.8
+
 ## 2.4.0-beta.7
 
 ### Minor Changes