+++ packages/generator-converters.ontologyir/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @osdk/generator-converters.ontologyir
 
+## 2.4.0-beta.9
+
+### Patch Changes
+
+- @osdk/client.unstable@2.4.0-beta.9
+
+## 2.4.0-beta.8
+
+### Patch Changes
+
+- @osdk/client.unstable@2.4.0-beta.8
+
 ## 2.4.0-beta.7
 
 ### Patch Changes