+++ packages/tool.generate-with-mock-ontology/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @osdk/tool.generate-with-mock-ontology
 
+## 0.4.0-beta.9
+
+### Patch Changes
+
+- Updated dependencies [9101bad]
+  - @osdk/client@2.4.0-beta.9
+  - @osdk/api@2.4.0-beta.9
+
+## 0.4.0-beta.8
+
+### Patch Changes
+
+- Updated dependencies [9f4fe9e]
+  - @osdk/client@2.4.0-beta.8
+  - @osdk/api@2.4.0-beta.8
+
 ## 0.4.0-beta.7
 
 ### Patch Changes